+++ src/screens/RecordEditor/NodeDefFormItem/NodeDefFormItem.tsx
@@ -40,7 +40,12 @@ export const NodeDefFormItem = (props: NodeComponentProps) => {
     parentNodeUuid,
     nodeDefUuid,
   });
+  const editable = DataEntrySelectors.useRecordNodePointerEditable({
+    parentNodeUuid,
+    nodeDefUuid,
+  });
   const keyAttributeLockAvailable =
+    editable &&
     NodeDefs.isKey(nodeDef) &&
     !NodeDefs.isEntity(nodeDef) &&
     !NodeDefs.isReadOnly(nodeDef) &&
@@ -57,7 +62,9 @@ export const NodeDefFormItem = (props: NodeComponentProps) => {
   const isLinkedToPreviousCycleRecord =
     DataEntrySelectors.useIsLinkedToPreviousCycleRecord();
   const canEditRecord = DataEntrySelectors.useCanEditRecord();
-  const alwaysVisible = Objects.isEmpty(NodeDefs.getApplicable(nodeDef));
+  const hasRelevantIf = Objects.isNotEmpty(NodeDefs.getApplicable(nodeDef));
+  const hasVisibleIf = Objects.isNotEmpty(NodeDefs.getVisibleIf(nodeDef));
+  const canDisappear = hasRelevantIf || hasVisibleIf;
 
   const includedInPreviousCycleLink =
     !NodeDefs.isKey(nodeDef) &&
@@ -98,8 +105,14 @@ export const NodeDefFormItem = (props: NodeComponentProps) => {
     [viewMode, styles.internalContainer],
   );
 
-  const editable =
-    canEditRecord && !NodeDefs.isReadOnly(nodeDef) && !keyAttributeLocked;
+  const editableFinal =
+    canEditRecord &&
+    !NodeDefs.isReadOnly(nodeDef) &&
+    editable &&
+    !keyAttributeLocked;
+
+  const showNodeComponentSwitch =
+    NodeDefs.isEntity(nodeDef) || editableFinal;
 
   const formItemComponent = useMemo(
     () => (
@@ -113,7 +126,7 @@ export const NodeDefFormItem = (props: NodeComponentProps) => {
           {isLinkedToPreviousCycleRecord && includedInPreviousCycleLink && (
             <PreviousCycleNodeValuePreview nodeDef={nodeDef} />
           )}
-          {editable ? (
+          {showNodeComponentSwitch ? (
             <NodeComponentSwitch
               nodeDef={nodeDef}
               parentNodeUuid={parentNodeUuid}
@@ -129,7 +142,7 @@ export const NodeDefFormItem = (props: NodeComponentProps) => {
       </VView>
     ),
     [
-      editable,
+      showNodeComponentSwitch,
       formItemComponentStyle,
       includedInPreviousCycleLink,
       internalContainerStyle,
@@ -141,13 +154,17 @@ export const NodeDefFormItem = (props: NodeComponentProps) => {
     ],
   );
 
-  if (alwaysVisible) {
-    return formItemComponent;
+  if (!visible) {
+    return null;
   }
 
-  if (settings.animationsEnabled && viewMode !== RecordEditViewMode.oneNode) {
+  if (
+    canDisappear &&
+    settings.animationsEnabled &&
+    viewMode !== RecordEditViewMode.oneNode
+  ) {
     return <Fade visible={visible}>{formItemComponent}</Fade>;
   }
 
-  return visible ? formItemComponent : null;
+  return formItemComponent;
 };