+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.6.0] - TO-BE-RELEASED
+
+### Added
+
+- Support "editable if" and "visible if"
+- Left side drawer: layout improvements;
+
 ## [2.5.9] - 2026-05-23
 
 ### Fixed